moved footer into a seperate component file; and now also shows up on login page.
